+++ husky_components/auto-complete/main.js
@@ -365,8 +365,8 @@ define([], function() {
                     url: this.options.prefetchUrl,
                     ttl: 1,
                     filter: function(data) {
-                        this.sandbox.emit(PREFETCH_LOAD.call(this));
-                        this.handleData(data);
+                        this.sandbox.emit(PREFETCH_LOAD.call(this), this.handleData(data));
+
                         return this.data;
                     }.bind(this)
                 };
@@ -382,10 +382,10 @@ define([], function() {
                         }.bind(this)
                     },
                     filter: function(data) {
-                        this.sandbox.emit(REMOTE_RETRIEVE.call(this));
+                        this.sandbox.emit(REMOTE_RETRIEVE.call(this), this.handleData(data));
                         this.hideLoader();
 
-                        return this.handleData(data);
+                        return this.data;
                     }.bind(this)
                 };
             }