+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG for husky
 
 * dev-master
+    * ENHANCEMENT #606 Added data to retrieve event
     * ENHANCEMENT #605 Added loader to indicate loading suggestions
 
 * 0.18.5 (2016-02-05)