+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG for husky
 
-* dev-master
+* 0.18.6 (2016-03-07)
     * HOTFIX      #609 Fixed input back class overlapping
     * HOTFIX      #607 Fixed data-attribute id for auto-complete
     * ENHANCEMENT #606 Added data to retrieve event